* xdisp.c (string_from_display_spec): Don't assume vecsize fits in int.
* indent.c (Fvertical_motion): Fix == vs = typo.

+++ src/xdisp.c
@@ -1226,7 +1226,7 @@ string_from_display_spec (Lisp_Object spec)
     }
   else if (VECTORP (spec))
     {
-      int i;
+      ptrdiff_t i;
 
       for (i = 0; i < ASIZE (spec); i++)
 	{
